requirements.txt
pydantic>=2.0.0
asyncpg>=0.29.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
google-genai>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
